When selecting a solution word for a date, wrap the index based on the length of the solution word array.
On 2021-11-18, `getWordIndex` returns 939. Both possibleWords.normal and
possibleWords.hard have 938 entries. This results in an
index-out-of-bounds exception, which returns `undefined` as the soltion.
This change wraps the index by the length of the possibleWords arrays,
so in this case, a `getWordIndex` of 939 will be wrapped to 0.

Note, it also appears that the first 3 entries in possibleWords.normal
were empty strings. I've removed those to prevent them for being used as
solutions but I am not sure if there are implications elsewhere if
possibleWords.normal for possibleWords.hard are of different lengths
(which they are after this change).